requirement.txt
Flask==3.0.3
pymongo==4.8.0
python-dotenv==1.0.1
PyJWT==2.8.0
bcrypt==4.1.3